+++ packages/remirror__extension-list/src/index.ts
@@ -1,7 +1,7 @@
 export { BulletListExtension } from './bullet-list-extension';
 export { toggleList } from './list-commands';
-export { ListItemSharedExtension } from './list-item-command-extension';
 export { ListItemExtension } from './list-item-extension';
+export { ListItemSharedExtension } from './list-item-shared-extension';
 export { OrderedListExtension } from './ordered-list-extension';
 export { TaskListExtension } from './task-list-extension';
 export { TaskListItemExtension } from './task-list-item-extension';
